refactor(chunk-store): Phase 4b — move staging bodies to chunk_staging.c
Moves all ChunkStaging-touching functions and their helpers from
chunk_store.c to chunk_staging.c, completing the staging
subsystem extraction begun in #880.

Function bodies moved (~225 lines):
  - Hash-table helpers (static within chunk_staging.c):
      csPendBucket, csPendHTRebuild, csPendHTEnsure, csRecentHTEnsure
  - Exposed in chunk_staging.h (called from chunk_store.c):
      csPendHTClear, csRecentHTClear
      csSearchPending, csSearchRecent
      csGrowPending, csGrowRecent, csGrowWriteBuf

Constants moved with the bodies:
  CS_PEND_HT_INIT_BITS, CS_PEND_HT_MAX_LOAD,
  CS_INIT_PENDING_ALLOC, CS_INIT_WRITEBUF_SIZE

chunk_staging.c now includes chunk_store.h to access the
ChunkStore struct that these functions operate on.

Verified:
  - make doltlite-lib + make doltlite (clean builds)
  - make lint (layering check passes)
  - Smoke: CREATE + dolt_commit + reopen + dolt_log count = 2
  - WAL regression tests (37/37):
      truncated_wal_rejected (6/6)
      wal_offset_corruption (5/5)
      reload_refs_transactional (13/13)
      refresh_refs_corruption_preserves_state (13/13)
  - Full C regression suite: 108622/108622 pass
